Add CLI module for command-line interface handling
- Introduced a new `cli.rs` file to encapsulate command-line argument parsing and routing.
- Implemented `Cli` struct with subcommands for Base64 and N-Triangle functionalities.
- Refactored `main.rs` to utilize the new CLI module, simplifying the main function and improving code organization.
- Integrated logging functionality based on verbosity flag.
