+++ src/cli.rs
@@ -1,0 +1,45 @@
+use std::env;
+
+use clap::{Parser, Subcommand};
+use env_logger;
+use log::Level;
+
+use crate::{base64, n_triangle};
+
+pub trait CliRoute {
+    fn route(&self);
+}
+
+#[derive(Parser, Debug)]
+#[command(author, version, about = "Algorithm tool", long_about = None)]
+pub struct Cli {
+    #[arg(short, long, help = "Show logs")]
+    pub verbose: bool,
+
+    #[command(subcommand)]
+    command: Commands,
+}
+
+#[derive(Subcommand, Debug)]
+enum Commands {
+    Base64(base64::Base64),
+    NTriangle(n_triangle::NTriangle),
+}
+
+impl CliRoute for Cli {
+    fn route(&self) {
+        if self.verbose {
+            env::set_var("RUST_LOG", Level::Trace.to_string());
+        }
+        env_logger::init();
+
+        match &self.command {
+            Commands::Base64(cli) => {
+                base64::exec(cli);
+            }
+            Commands::NTriangle(cli) => {
+                n_triangle::exec(cli);
+            }
+        }
+    }
+}
